AdminStarmap: Alarm Rot wird jetzt in der AdminStarmap wieder überall angezeigt.
